added unit test, integration test, sonarcloud

+++ tennis-supplies-backend/src/main/java/com/tennissupplies/tennissuppliesbackend/services/CategoryService.java
@@ -3,6 +3,7 @@ package com.tennissupplies.tennissuppliesbackend.services;
 import com.tennissupplies.tennissuppliesbackend.controller.WebSocketController;
 import com.tennissupplies.tennissuppliesbackend.models.Category;
 import com.tennissupplies.tennissuppliesbackend.repository.CategoryRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -23,16 +24,23 @@ public class CategoryService {
     }
 
     public Category getCategoryById(Long id) {
-        return categoryRepository.findById(id).orElse(null);
+        return categoryRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Category not found with id: " + id));
     }
 
     public Category saveCategory(Category category) {
+        if (category == null || category.getName() == null || category.getName().isEmpty()) {
+            throw new IllegalArgumentException("Category name cannot be null or empty");
+        }
         Category cat = categoryRepository.save(category);
         webSocketController.notifyProductChange();
         return cat;
     }
 
     public void deleteCategory(Long id) {
+        if (!categoryRepository.existsById(id)) {
+            throw new EntityNotFoundException("Category with id " + id + " not found");
+        }
         categoryRepository.deleteById(id);
         webSocketController.notifyProductChange();
